Create delete location method

+++ src/main/java/com/sasha/buildland/repository/LocationRepository.java
@@ -3,5 +3,5 @@ package com.sasha.buildland.repository;
 import com.sasha.buildland.entity.Location;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LocationRepository extends CrudRepository<Location, Integer> {
+public interface LocationRepository extends CrudRepository<Location, Long> {
 }
